Fix bug with out-of-time production switches
Problem:
Switching production didn't work under some conditions.

Cause:
In getPlanData, the if statement controlling flow to the block
for processing switch actions was performing two checks: that this
action is a switch action and that it was within time.

When a switch action wasn't in time it would fail that check and
incorrectly flow into upgrade processing. Since it wasn't an
upgrade action, this failed and the function returned null. If
getPlanData returns null, the state is not updated. Hence, the
switch update "not working".

For some reason, this bug only showed itself when the user adjusted
production such that there was 0 production of a needed resource.