Add desktop mode formatting
Switches to two-column layout.
Forces toggles to "on" for game status and offline periods and hides
the toggle buttons.
Spreads out the planner so switches have their own column and timegroup
headings are wider.
Added extra breakpoint, since the lower end of the md breakpoint
wasn't wide enough for both the two-column layout and the spread out
planner.