Adjust top banner content; add Tooltip
Reworded top banner to emphasise that win/lose is about 1st prize.
Reworded top banner to describe the second number as "rec." rather
than "best", which is a bit less misleading.
Added a tooltip and a ? button to abbreviate to fit on mobile and
look more consistent with the "now" number.