Adjust game state display based on mode
When the user has selected "plan mode" and entered only a start time
and premium info, it's very confusing to confront them with a panel
displaying the full game state, because that includes several irrelevant
in-progress variables which they didn't enter and can't alter.

It's also a little confusing for users who picked "active mode" to have
a start time variable in the middle of their information, when they have
no means of directly altering this.

To resolve that confusion, the game state display panel now limits its
fields to those relevant to the currently selected input mode.

Related small adjustments:
* Amended usePlan to return the actual mode instead of "isActive" and
  "isPlan". I'd been hoping to avoid importing the enum everywhere, but
  it felt messy
* Altered the width of the labels in the game state display so that the
  displayed values would line up with the end of the edit button,
  improving the appearance in plan mode
* Moved custom grid CSS from styles.css into the className string: now
  that I know you can do that, I'd rather have all the CSS in one place
* Adjusted appearance of dust stockpile
* In the timeGroup's "more" panel, updated the appearance of the dust
  stockpile to match the new appearance in game state display
* In the timeGroup's "more" panel, changed the dust label from
  "Stockpiled" to "stockpile" for consistency with the eggs table
* Renamed some labels for clarity