test: cover missing variables in sectionGameState
Tests for levels and stockpiles in sectionGameState were only testing
one example in each category, which left open the possibility of a
problem with one of the other variables in those categories. Refactored
it to test all levels and stockpiles.

In the process I learned of Jest's "test.each(...)" feature.
formActiveMode also had several tests that would benefit from
"test.each(...)", so those have been refactored for DRYness.