feat and fix: top switch button
Purpose of Feature:
The switch button at the top works differently to the other switch
buttons: it alters production "now" instead of "after purchasing the
previous upgrade". Its purpose is to allow users to make adjustments
during production of the first unpurchased planned upgrade.

If the game status hasn't been updated in a while, it's possible for
"now" to be a later time than the first planned purchase. In this case
the top switch button does absolutely nothing.

It's very confusing for it to allow users to open it, make selections,
submit the form and then do nothing. This feature addresses that.

Feature:
When opened by the top switch, the production settings modal now
displays a message.
* If "now" is a good time, it informs the user that they are changing
  the production settings starting from the current time, which updates
  while the modal is open
* If "now" is a bad time, it displays a warning message and disables
  the entire form
* If "now" changes from good to bad while the modal is open, the good
  message is replaced by the bad and the form is disabled

---
Bugs Fixed
* Results at top: if the last action in actions[] was a switch action,
  it was being ignored in results calculations, due to the way
  production settings at the end were calculated
* Switching production was sometimes behaving oddly: desired changes
  were discarded, unwanted ones were added. Issue was caused by unwanted
  switchData being added to the timeGroup of the first purchased
  upgrade under some circumstances