fix/feat: improved handling of zero production
Bug #1:
If the top production switch was used to adjust production and this
resulted in zero production of an egg needed to purchase the first
upgrade, the entire program would crash.

Cause:
If the first upgrade can't be purchased, there are no valid time groups.
Some results calculations are based on there being at least one valid
time group: they had a check for /zero/ time groups, but not for
"one, but invalid",

Feature:
Fixing this raised the question: what /should/ users see in this case,
apart from the insert and switch buttons they'd need to fix it?
Added a message which explains which eggs are needed for the next
upgrade and which of those have zero production. Since this message
would also be useful if zero production occurred at the end of the plan,
I set it up to work in both cases.

Bug #2:
The top production settings modal was initialising incorrectly when the
first upgrade was a new worker: the new worker, who hadn't been unlocked
yet and so should be disabled, was instead showing as available.

Small changes in support of these changes:
* Badge wrapper is now a span instead of a div so it can be used within
  a sentence span
* Added CSS to prevent ControlsRow's buttons from collapsing together
* While investigating the "zero production crash", I decided that
  Future!Me would not remember why I had half the handleSwitchAction
  work inside a block conditional on flagFirstisDone. What started as a
  more detailed comment ended with breaking up the function a little