fix: offline period validation and display
Bug:
Entering an offline period that spanned the entire event caused an
empty plan with no indication of the cause of the problem.

Rejected:
Considered displaying an error message instead of an empty plan, but
that would require a means to identify when the error is caused by
offline periods. Doing so would require accounting for the following
properties of offline periods:
* They can be of a longer duration than the event
* They can overlap
* They can be duplicated
* There can be any number of them

Conclusion:
Offline period form validation is too permissive. Better to prevent
users from entering an offline period that would cause this error in
the first place.

Changes:
* Form for offline periods has gained some extra validation rules. In
  addition to previous rules, there is now: a maximum of 150 periods;
  no overlapping; no duplicates; no being offline for the entire event
* Display of offline periods now shows them in order, grouped by date
* Edit modal now identifies the target offline period by its previous
  start and end times
* "Position numbers", previously used to identify the target offline
  period in the edit modal, have been removed
* Overflow-Y scrollbar added to the container with the dates and
  offline periods, so the heading and add button remain visible
* calcDateDisplayStr has been renamed to calcDateWithTimeDisplayStr, to
  more accurately reflect its output
* Date and time can now be formatted separately using the new
  calcDateDisplayStr and calcTimeDisplayStr functions