fix: NaN error in time group more
Bug:
The time group's "more" panel was sometimes displaying "NaN NaN:NaN"
instead of a finish time, "done" or "-".

Cause:
The function calculating the time ID when production would be complete
was checking for `current > goal`. When goal and current were both 0,
the logical test would fail, so it would proceed to the next step:
dividing 0 by the production rate, hence NaN.

`current >= goal` means it correctly returns 0, which the "more" panel
displays correctly.