fix: time entered (and time remaining)
Problem:
Delays at any point in the form-filling process could throw off the
link between "timeEntered" and the time at which the entered values for
time remaining, stockpiles and levels were correct, making a nonsense of
offline periods and estimated purchase times.

The previous solution - displaying "time entered" and a button to update
it to "now" - was inadequate. How would the user know that "time
entered" actually means "time at which these values were correct"?
What if they'd intentionally entered slightly old data and needed an
/earlier/ time?

Solution:
Replaced the old form inputs with a fieldset labelled "Status At" which
contains selects for the hour, minute and - to support the date rolling
over - "today or yesterday".

* Initialises to "now", so players promptly entering current data lose
  nothing, they can skip over the fieldset as they skipped over the
  button
* Players wishing to enter older data can now do so (within 24hrs).
* For unintentional delays, hopefully the fieldset better conveys its
  purpose and stands out a bit more

Also:
* timeEntered renamed to "timestamp" in the code to more accurately
  convey its purpose
* Refactored the hook for the active game status form to pass out a
  closure for setting timeRemaining instead of the set state action(*)
  and removed a closure used only by the old timeEntered